after refreshing the page vehicle info is coming
