+++ src/app/main/apps/vehicles/VehicleInfo.js
@@ -1,39 +1,32 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import { getVehicle, vehicleInfo } from './store/vehiclesSlice';
+import withReducer from 'app/store/withReducer';
+import { getVehicle, vehicleInfo } from './store/vehicleSlice';
 import VehicleMaintanceTable from './VehicleMaintanceTable';
+import reducer from './store';
+
 import '../../../../styles/vehicle-info.css';
-// import BasicTable from './VehicleMaintanceInfo';
 
 function VehicleInfo() {
-  // const vehicles = useSelector(state => state.vehiclesApp.vehicles.vehicleDialog.selectedRow);
   const vehicles = useSelector(vehicleInfo);
 
   const dispatch = useDispatch();
   const { id } = useParams();
 
-  // const store = useSelector(state => state.vehiclesApp);
-  // const selectedVehicle = useSelector(({ vehiclesApp }) => vehiclesApp.vehicles.selectedVehicle);
-
-  // const selectedVehicle = useSelector(selectVehiclesById);
-
   useEffect(() => {
     dispatch(getVehicle({ id }));
-  }, [dispatch, id]);
-  // let status = vehicles.active;
-
-  // if (!vehicles) {
-  //   return <div>Not found</div>;
-  // }
+  }, []);
 
   return (
     <>
+      <div className="header">
+        <p>Vehicles details</p>
+      </div>
       {vehicles ? (
         <div className="main-container-info">
           <div className="container-info">
             <div>
-              <h1>Vehicles details</h1>
               <h3 className="text-info">Status: {vehicles.active ? 'Available' : 'Unavailable'}</h3>
               <h3 className="text-info">Plate number: {vehicles.plate_number}</h3>
               <h3 className="text-info">Model: {vehicles.model}</h3>
@@ -43,7 +36,7 @@ function VehicleInfo() {
             </div>
 
             <div className="img-info">
-              <img src={vehicles.image_url} alt="no image"></img>
+              <img src={vehicles.image_url} alt="no found" />
             </div>
           </div>
           <div>
@@ -52,9 +45,9 @@ function VehicleInfo() {
           </div>
         </div>
       ) : (
-        <div>Not found</div>
+        <div>Loading...</div>
       )}
     </>
   );
 }
-export default VehicleInfo;
+export default withReducer('vehiclesApp', reducer)(VehicleInfo);
