+++ src/app/main/apps/vehicles/store/index.js
@@ -1,9 +1,11 @@
 import { combineReducers } from '@reduxjs/toolkit';
 import vehicles from './vehiclesSlice';
+import vehicle from './vehicleSlice';
 import user from './userSlice';
 
 const reducer = combineReducers({
   vehicles,
+  vehicle,
   user
 });
