+++ src/app/main/apps/vehicles/store/vehiclesSlice.js
@@ -13,14 +13,7 @@ export const getVehicles = createAsyncThunk('vehicle-list-app/vehicles/getVehicl
   });
   return response.data;
 });
-export const getVehicle = createAsyncThunk('vehiclesApp/vehicles/getVehicle', async ({ id }) => {
-  const response = await axios.get(`${VEHICLES_API}/${id}`, {
-    headers: {
-      Authorization: TOKEN
-    }
-  });
-  return response.data;
-});
+
 // export const getVehicles = createAsyncThunk(
 //   'vehicle-list-app/vehicles/getVehicles',
 //   async (routeParams, { getState }) => {
@@ -166,7 +159,6 @@ const vehiclesSlice = createSlice({
       },
       data: null
     },
-    selectedVehicle: null,
     confirmDialog: {
       type: 'delete',
       props: {
@@ -254,16 +246,9 @@ const vehiclesSlice = createSlice({
       vehiclesAdapter.setAll(state, data);
       state.routeParams = routeParams;
       state.searchText = '';
-    },
-    [getVehicle.fulfilled]: (state, action) => {
-      // vehiclesAdapter.setAll(state, data);
-      state.routeParams = action.payload.routeParams;
-
-      state.selectedVehicle = action.payload;
     }
   }
 });
-export const vehicleInfo = state => state.vehiclesApp?.vehicles.selectedVehicle;
 
 export const {
   setVehiclesSearchText,
